% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_topSNPs_sub.R
\name{auto_topSNPs_sub}
\alias{auto_topSNPs_sub}
\title{Automatically identify top SNP per locus}
\usage{
auto_topSNPs_sub(topSNPs, query, locus)
}
\arguments{
\item{topSNPs}{Output of \link[echodata]{import_topSNPs}.}

\item{query}{Summary stats subset.}

\item{locus}{Locus name.}
}
\description{
If no \code{topSNPs} dataframe is supplied,
 this function will sort by p-value and then effect size,
 and use the SNP in the first row.
}
\seealso{
Other standardization functions: 
\code{\link{calculate_tstat}()}
}
\concept{standardization functions}
\keyword{internal}
