% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart_geneInfo.R
\name{biomart_geneInfo}
\alias{biomart_geneInfo}
\title{Get gene info using Biomart}
\source{
\href{https://bioconductor.org/packages/release/bioc/html/biomaRt.html}{
biomaRt}
}
\usage{
biomart_geneInfo(geneList, reference_genome = "grch37")
}
\description{
Get gene info using Biomart
}
\examples{
gene_info <- echodata::biomart_geneInfo(geneList = c("PTK2B", "CLU", "APOE"))
}
\concept{annotate}
\keyword{internal}
