% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BST1}
\alias{BST1}
\title{\pkg{echolocatoR} output example: BST1 locus}
\format{
data.table
\describe{
  \item{SNP}{SNP RSID}
  \item{CHR}{Chromosome}
  \item{POS}{Genomic position (in basepairs)}
  \item{...}{Optional: extra columns}
}
}
\source{
\url{https://doi.org/10.1016/S1474-4422(19)30320-5}
}
\usage{
BST1
}
\description{
An example results file after running
\code{\link[echolocatoR]{finemap_loci()}}.
}
\details{
Data originally comes from the Parkinson's disease GWAS
by \href{https://doi.org/10.1016/S1474-4422(19)30320-5}{
Nalls et al. (The Lancet Neurology)}.
}
\examples{
\dontrun{
BST1 <- echolocatoR::BST1
usethis::use_data(BST1, overwrite = TRUE)
}
}
\keyword{datasets}
