% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BST1_LD_matrix}
\alias{BST1_LD_matrix}
\title{LD with the lead SNP: BST1 locus}
\format{
data.table
\describe{
  \item{SNP}{SNP RSID}
  \item{CHR}{Chromosome}
  \item{POS}{Genomic position (in basepairs)}
  \item{...}{Optional: extra columns}
}
}
\source{
\url{https://www.ukbiobank.ac.uk}
\url{https://doi.org/10.1038/s41588-020-00735-5}
 @examples
 \code{ 
data("BST1")
# Only including a small subset of the full
# LD matrix for storage purposes.
lead_snp <- subset(BST1, leadSNP)$SNP
snp_list <-  BST1[which(BST1$SNP==lead_snp)-
                      seq(100,which(BST1$SNP==lead_snp))+100,]$SNP
path <- portal_query(phenotypes = "parkinson", 
                     loci = "BST1", 
                     dataset_types = "LD", 
                     LD_panels = "UKB")
# path <- file.path(
#     "../Fine_Mapping/Data/GWAS/Nalls23andMe_2019/BST1/plink/UKB_LD.RDS")
BST1_LD_matrix <- readRDS(path)
BST1_LD_matrix <- BST1_LD_matrix[snp_list, snp_list]
### shortcut
# BST1_LD_matrix <- echolocatoR::BST1_LD_matrix
usethis::use_data(BST1_LD_matrix, overwrite=TRUE)
 }
}
\usage{
data("BST1_LD_matrix")
}
\description{
Precomputed LD within the \emph{BST1} locus
 (defined in \code{\link[=BST1]{BST1}}.
LD derived white British subpopulation in the UK Biobank.
Only includes a subset of all the SNPs for storage purposes
(including the lead GWAS/QTL SNP).
}
\details{
Data originally comes from \href{https://www.ukbiobank.ac.uk}{UK Biobank}.
LD was pre-computed and stored by the Alkes Price lab
(see \href{https://doi.org/10.1038/s41588-020-00735-5}{here}).
}
\keyword{datasets}
