% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_executable.R
\name{check_executable}
\alias{check_executable}
\title{Check executable}
\usage{
check_executable(path, verbose = TRUE)
}
\arguments{
\item{path}{Path or named list of paths to executable(s).}

\item{verbose}{Print messages.}
}
\value{
Logical list.
}
\description{
Check that an executable works.
}
\keyword{internal}
