% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/column_dictionary.R
\name{column_dictionary}
\alias{column_dictionary}
\title{Map column names to positions.}
\usage{
column_dictionary(path)
}
\arguments{
\item{path}{Path to full summary stats file
(or any really file you want to make a column dictionary for).}
}
\value{
Named list of column positions.
}
\description{
Useful in situations where you need to specify columns by index 
instead of name (e.g. awk queries).
}
\examples{
tmp <- tempfile()
data.table::fwrite(echodata::BST1, tmp)
cdict <- echodata::column_dictionary(path = tmp)
}
