% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct_colmap.R
\name{construct_colmap}
\alias{construct_colmap}
\title{Target column map}
\usage{
construct_colmap(
  munged = FALSE,
  CHR = "CHR",
  POS = "POS",
  SNP = "SNP",
  P = "P",
  Effect = "Effect",
  StdErr = "StdErr",
  tstat = "tstat",
  Locus = "Locus",
  Freq = "Freq",
  MAF = "MAF",
  A1 = "A1",
  A2 = "A2",
  Gene = "Gene",
  N_cases = "N_cases",
  N_controls = "N_controls",
  proportion_cases = "calculate",
  sample_size = NULL,
  verbose = TRUE
)
}
\arguments{
\item{munged}{Whether \code{fullSS_path} have already been
standardised/filtered  full summary stats
with \link[MungeSumstats]{format_sumstats}.
If \code{munged=FALSE} you'll need to provide the necessary
 column names to the \code{colmap} argument.}

\item{CHR}{Name of the chromosome column in the full summary stats
file.
Can be "chr1" or "1" format.
(\emph{default: ="CHR"})}

\item{POS}{Name of the genomic position column in the full summary
stats file.
Must be in units of basepairs.
(\emph{default: ="POS"})}

\item{SNP}{Name of the SNP RSID column in the full summary stats file.
(\emph{default: ="SNP"})}

\item{P}{Name of the p-value column in the full summary stats file.
Raw p-values are preferred, but if not available corrected p-values
(e.g. FDR) can be used instead.
(\emph{default: ="P"})}

\item{Effect}{Name of the effect size column in the full summary stats file.
Effect size is preferred, but if not available other metrics like Beta for
Odds Ratio can be used instead.
(\emph{default: ="Effect"})}

\item{StdErr}{Name of the standard error
column in the full summary stats file.
You can also set \code{stderr="calculate"} to infer standard error
using: \code{effect / tstat}.
(\emph{default: ="StdErr"})}

\item{tstat}{Name of the t-statistic column in the full
summary stats file.
This column is not necessary unless \code{stderr="calculate"}
or the standard error column is missing.
(\emph{default: ="t-stat"})}

\item{Locus}{Name of the locus column in the full summary stats file.
(\emph{default: ="Locus"})}

\item{Freq}{Name of the allele frequency column in the full
 summary stats file.
Effect allele frequency is preferred, but the non-effect allele can
be provided instead (though this may be less accurate).
This column is not necessary unless \code{MAF="calculate"} or
 the MAF column is missing.
(\emph{default: ="Freq"})}

\item{MAF}{Name of the minor allele frequency column in the
full summary stats file.
Can be inferred from \strong{freq} if missing from the dataset.
(\emph{default: ="MAF"})}

\item{A1}{Name of the effect/risk allele column in the full
summary stats.}

\item{A2}{Name of the other/non-risk allele column in the full
summary stats.}

\item{Gene}{For QTL studies, the name of the \[e\]gene column in the
full summary stats file (\emph{default: "gene"}).
This column will be used for filtering summary stats if supplying a named
list of gene:Locus pairs to \code{loci}.}

\item{N_cases}{The number of case subjects in the study.
Instead of providing a redundant \strong{N_cases} column,
you can simply enter one value here.}

\item{N_controls}{The number of control subjects in the study.
Instead of providing a redundant \strong{N_controls} column,
you can simply enter one value here.}

\item{proportion_cases}{The proportion of total subjects in the
study that were cases.
 if \code{proportion_cases="calculate"} then this is inferred:
 \code{N_controls / N_controls}.}

\item{sample_size}{The overall sample size of the study.
If none is given, and \strong{N_cases} and \strong{N_controls}
columns are present,
then sample_size is inferred to be: \code{max(N_cases) + max(N_controls)}.}

\item{verbose}{Print messages.}
}
\description{
This function wraps all column name mappings for the full
GWAS/QTL summary statistics file specified by \code{fullSS_path}
 (i.e. the "target").
}
\examples{
colmap <- echodata::construct_colmap(munged=TRUE)
}
