% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_to_rsids.R
\name{coords_to_rsids}
\alias{coords_to_rsids}
\title{Get RSIDS from SNP coordinates}
\usage{
coords_to_rsids(
  dat,
  genome_build = "hg19",
  drop_unannotated = TRUE,
  drop_duplicates = TRUE,
  snp_colname = "SNP",
  verbose = TRUE
)
}
\arguments{
\item{dat}{Data.table of SNP positions,
with the columns \code{CHR} and \code{POS}. Extra columns allowed.}

\item{genome_build}{Which genome build \code{dat} is in
(e.g. "hg19" or "hg38").}

\item{drop_unannotated}{Drop SNPs that RSIDs couldn't be found for.}

\item{drop_duplicates}{Drop any duplicate SNPs rows.}
}
\description{
Get RSIDS from SNP coordinates using Bioconductor genome references.
}
\examples{
dat <- dplyr::rename(echodata::BST1, SNP_old=SNP)
dat_annot <- echodata::coords_to_rsids(dat = dat)
}
