% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDT.R
\name{createDT}
\alias{createDT}
\title{Interactive DT}
\usage{
createDT(DF, caption = "", scrollY = 400)
}
\arguments{
\item{DF}{Data to show.}

\item{caption}{Table caption.}

\item{scrollY}{Maximum height (in pixels) before defaulting to scrolling.}
}
\description{
Generate an interactive data table with download buttons.
}
\seealso{
Other general: 
\code{\link{createDT_html}()}
}
\concept{general}
