% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_genes.R
\name{detect_genes}
\alias{detect_genes}
\title{Detect QTL genes in full summary stats file}
\usage{
detect_genes(loci, verbose = FALSE)
}
\description{
Allows summary stats from different genes to be
fine-mapped separately.
}
\examples{
#### No genes (names) ####
loci <- c("BST1","LRKR2","MEX3C")
detect_genes(loci)

#### Genes (names) ####
loci <- c(BST1="BST1", LRRK2="LRRK2", MEX3C="MEX3C")
echodata::detect_genes(loci)
}
