% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_fullSS.R
\name{example_fullSS}
\alias{example_fullSS}
\title{Store \code{fullSS_dat}}
\usage{
example_fullSS(
  save_dir = tempdir(),
  dataset = c("Nalls2019", "Kunkle2019"),
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{save_dir}{Path to save the example summary statistics to.
It is usually best to provide the absolute path
 rather than the relative path.}

\item{dataset}{Name of the example GWAS dataset to use.}

\item{nThread}{Number of threads to use.}

\item{verbose}{Print messages.}
}
\value{
Path to saved summary statistics file.
}
\description{
Save downsampled genome-wide summary statistics from different GWAS to disk.
}
\details{
\itemize{
\item{"Nalls2019"}{Parkinson's Disease GWAS from
\href{https://doi.org/10.1016/S1474-4422(19)30320-5}{
Nalls et al. 2019} (excluding 23andMe data).}
\item{"Kunkle2019"}{Alzheimer's Disease GWAS from
\href{https://www.nature.com/articles/s41588-019-0358-2}{
Kunkle et al. 2019}.}
}
}
\examples{
fullSS_path <- example_fullSS()
}
