% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_loci.R
\name{extract_loci}
\alias{extract_loci}
\title{Extract loci from GWAS/QTL summary statistics}
\usage{
extract_loci(topSNPs, fullSS, bp_distance = 5e+05, munged = TRUE)
}
\arguments{
\item{topSNPs}{Positions of the top SNPs prepare using
\link[echodata]{import_topSNPs}.}

\item{fullSS}{\link[data.table]{data.table} of the 
full GWAS/QTL summary statistics.}

\item{bp_distance}{The basepair distance upstream/downstream of the proxy 
SNP (in \code{topSNPs}) to extract.}

\item{munged}{Whether \code{fullSS} was previously munged with 
\link[MungeSumstats]{format_sumstats}.}
}
\description{
Extract loci from GWAS/QTL summary statistics
}
