% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_snps.R
\name{filter_snps}
\alias{filter_snps}
\title{Filter SNPs}
\usage{
filter_snps(
  dat,
  bp_distance = 5e+05,
  remove_variants = FALSE,
  min_POS = NA,
  max_POS = NA,
  max_snps = NULL,
  min_MAF = NULL,
  trim_gene_limits = FALSE,
  verbose = TRUE
)
}
\description{
Filter SNPs by MAF, window size, min/max position, maximum number of SNPs,
 or gene coordinates.
You can also explicitly remove certain variants.
}
\examples{
dat <- echodata::filter_snps(dat = echodata::BST1)
}
\seealso{
Other SNP filters: 
\code{\link{gene_trimmer}()},
\code{\link{limit_snps}()}
}
\concept{SNP filters}
