% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_consensus_snps.R
\name{find_consensus_snps}
\alias{find_consensus_snps}
\title{Find Consensus SNPs}
\usage{
find_consensus_snps(
  dat,
  credset_thresh = 0.95,
  consensus_thresh = 2,
  sort_by_support = TRUE,
  exclude_methods = NULL,
  verbose = TRUE
)
}
\arguments{
\item{dat}{Fine-mapping results data.}

\item{credset_thresh}{The minimum mean Posterior Probability
(across all fine-mapping methods used) of SNPs to be included in 
the "mean.CS" column.}

\item{consensus_thresh}{The minimum number of fine-mapping tools 
in which a SNP is in the Credible Set in order to be included in the 
"Consensus_SNP" column.}

\item{sort_by_support}{Whether to sort SNPs by the "Support" column.}

\item{exclude_methods}{Whether to exclude results from 
certain fine-mapping methods when mean/consensus columns.}

\item{verbose}{Print messages.}
}
\description{
Find Consensus SNPs in \link[echolocatoR]{finemap_loci} output.
}
\examples{
BST1 <- echodata::find_consensus_snps(dat = echodata::BST1)
}
