% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_consensus_snps_no_polyfun.R
\name{find_consensus_snps_no_polyfun}
\alias{find_consensus_snps_no_polyfun}
\title{Find Consensus SNPs: without PolyFun}
\usage{
find_consensus_snps_no_polyfun(dat, verbose = TRUE)
}
\arguments{
\item{dat}{Fine-mapping results data.}

\item{verbose}{Print messages.}
}
\description{
Find Consensus SNPs in \link[echolocatoR]{finemap_loci} output, without 
using any results from the functional fine-mapping tool PolyFun 
(in a new column "Consensus_SNP_noPF"). 
This helps to avoid any circularity when validating results with functional 
annotation that PolyFun may have been trained on.
}
\examples{
BST1 <- echodata::find_consensus_snps_no_polyfun(dat = echodata::BST1)
}
\concept{echolocatoR}
