% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_executable.R
\name{find_executable}
\alias{find_executable}
\title{Find executable}
\usage{
find_executable(
  path = NULL,
  tool = c("bcftools", "gcta", "plink"),
  verbose = TRUE
)
}
\arguments{
\item{path}{User-provided path to executable. 
Is \code{NULL} (default), an executable will be downloaded 
and stored automatically.}

\item{verbose}{Print messages.}
}
\value{
Path to executable.
}
\description{
Find the path to an executable binary for a given software.
}
\examples{
paths <- find_executable()
}
\keyword{internal}
