% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_top_consensus.R
\name{find_top_consensus}
\alias{find_top_consensus}
\title{Find the top Consensus SNP}
\usage{
find_top_consensus(dat, n_top = 1, grouping_vars = c("Locus"))
}
\arguments{
\item{dat}{Fine-mapping results data.}

\item{n_top}{Number of top SNPs per locus to include.}

\item{grouping_vars}{Variables to group by when selecting top N rows.}
}
\description{
Identify the \code{n_top} Consensus SNP(s) per Locus,
defined as the Consensus SNPs with the highest mean PP
across all fine-mapping tools used.
}
\examples{
dat <- echodata::find_top_consensus(dat = echodata::BST1)
}
