% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_coltypes.R
\name{fix_coltypes}
\alias{fix_coltypes}
\title{Automatically fix column types}
\usage{
fix_coltypes(dat)
}
\arguments{
\item{dat}{A data.frame.}
}
\value{
A fixed data.frame
}
\description{
Sometimes things that should be numbers get interpreted as strings
 (e.g. 1 vs. "1"). This function automatically inferred the best match 
 for each column and reformats its column type accordingly.
}
\examples{
dat <- mtcars
dat$cyl <- as.character(dat$cyl)
dat$wt <- as.character(dat$wt)
dat$drat <- as.factor(dat$drat)

dat2 <- fix_coltypes(dat=dat)
}
