% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_locus_list.R
\name{gene_locus_list}
\alias{gene_locus_list}
\title{Generate a named list of (e)Gene-Locus pairs}
\usage{
gene_locus_list(
  loci = NULL,
  topSNPs = NULL,
  dataset_type = NULL,
  verbose = TRUE
)
}
\arguments{
\item{loci}{User-supplied locus list.}

\item{topSNPs}{Output of the function \link[echodata]{import_topSNPs}, 
containing the top (index) SNPs per locus.}

\item{dataset_type}{The kind dataset you're fine-mapping
(e.g. GWAS, eQTL, tQTL).}

\item{verbose}{Print messages.}
}
\value{
Named list
}
\description{
This object is to be passed to the \code{loci=} argument in 
\code{echolocatoR::finemap_loci()}.
}
\examples{
loci <- gene_locus_list(topSNPs = echodata::topSNPs_Nalls2019,
                        dataset_type = "QTL")
}
