% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CS_bins.R
\name{get_CS_bins}
\alias{get_CS_bins}
\title{Get Credible Set bins}
\usage{
get_CS_bins(merged_DT)
}
\arguments{
\item{merged_DT}{Merged data.table of fine-mapped summary statistics.}
}
\description{
Count bins of tool-specific and union Credible Set sizes.
}
\examples{
merged_DT <- echodata::get_Nalls2019_merged()
bin_counts <- get_CS_bins(merged_DT = merged_DT) 
}
\seealso{
Other summarise: 
\code{\link{get_CS_counts}()}
}
\concept{summarise}
