% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_CS_counts.R
\name{get_CS_counts}
\alias{get_CS_counts}
\title{Get Credible Set SNP counts}
\usage{
get_CS_counts(merged_DT, top_CS_only = FALSE)
}
\arguments{
\item{merged_DT}{Merged data.table of fine-mapped summary statistics.}

\item{top_CS_only}{If a given method contains multiple Credible Sets, 
only choose one.}
}
\description{
Tally tool-specific and Union Credible Set sizes
 across fine-mapping methods and across loci.
}
\examples{
merged_DT <- echodata::get_Nalls2019_merged()
locus_order <- get_CS_counts(merged_DT = merged_DT) 
}
\seealso{
Other summarise: 
\code{\link{get_CS_bins}()},
\code{\link{get_SNPgroup_counts}()},
\code{\link{results_report}()}
}
\concept{summarise}
