% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Kunkle2019_coloc.R
\name{get_Kunkle2019_coloc}
\alias{get_Kunkle2019_coloc}
\title{Get colocalization results with Kunkle2019 GWAS}
\source{
\code{
URL <- paste(
    "https://github.com/RajLabMSSM",
    "Fine_Mapping_Shiny/raw/master/www",
    "Microglia_all_regions_Kunkle_2019_COLOC.snp-level_select.tsv.gz",
     sep="/")
tmp <- file.path(tempdir(), basename(URL))
utils::download.file(URL, tmp)
piggyback::pb_upload(file = tmp, repo = "RajLabMSSM/echodata")
}
}
\usage{
get_Kunkle2019_coloc(return_path = FALSE)
}
\arguments{
\item{return_path}{If \code{TRUE}, return a named list containing both
the imported data and the path where it has been saved.
Default: \code{FALSE}.}
}
\value{
\link[data.table]{data.table}.
}
\description{
Get colocalization results computed by 
\href{https://doi.org/10.1038/s41588-021-00976-y}{Lopes et al. 2022}
between the \href{https://doi.org/10.1038/s41588-019-0358-25}{
Kunkle2019 Alzheimer's disease GWAS} and microglia eQTL datasets.
}
\examples{
coloc_res <- get_Kunkle2019_coloc()
}
