% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Nalls2019_coloc.R
\name{get_Nalls2019_coloc}
\alias{get_Nalls2019_coloc}
\title{Get colocalization results with Nalls2019 GWAS}
\source{
\code{
URL <- paste("https://github.com/RajLabMSSM/Fine_Mapping_Shiny/raw",
      "master/www/data/GWAS/Nalls23andMe_2019/_genome_wide/COLOC",
      "coloc.eQTL_Catalogue_ALL.csv.gz", sep="/")
tmp <- file.path(tempdir(), basename(URL))
utils::download.file(URL, tmp)
piggyback::pb_upload(file = tmp, repo = "RajLabMSSM/echodata")
}
}
\usage{
get_Nalls2019_coloc()
}
\description{
Get colocalization results computed by 
\href{https://doi.org/10.1093/hmg/ddab294}{Schilder et al. 2022}
between the \href{https://doi.org/10.1016/S1474-4422(19)30320-5}{
Nalls2019 Parkinson's disease GWAS} and all eQTL datasets in the 
\href{https://www.ebi.ac.uk/eqtl/}{eQTL Catalogue}.
}
\examples{
coloc_res <- get_Nalls2019_coloc()
}
