% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_Nalls2019_loci.R
\name{get_Nalls2019_loci}
\alias{get_Nalls2019_loci}
\title{\pkg{echolocatoR} output example: select loci}
\usage{
get_Nalls2019_loci(
  save_dir = tools::R_user_dir(package = "echodata", which = "cache"),
  return_paths = TRUE,
  return_dir = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{save_dir}{Local directory to cache data in.}

\item{return_paths}{Returns local paths to cached fine-mapping results
(default: \code{TRUE}), or the data itself as a named list (\code{FALSE}).}

\item{return_dir}{Return the directory name
instead of the individual file paths.}

\item{verbose}{Print messages.}
}
\value{
File paths
}
\description{
An example results file after running \code{finemap_loci}
 on 3 Parkinson's disease (PD)-associated loci (BST1, LRRK2, and MEX3C).
 
Data originally comes from the PD GWAS
by \href{https://doi.org/10.1016/S1474-4422(19)30320-5}{
Nalls et al. (The Lancet Neurology)}.
}
\examples{
files <- get_Nalls2019_loci()
}
