% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nrows.R
\name{get_nrows}
\alias{get_nrows}
\title{Get nrows}
\usage{
get_nrows(path)
}
\arguments{
\item{path}{Path to file saved on disk.}
}
\description{
Get the number of rows of a large file without reading it into memory.
}
\seealso{
Other general: 
\code{\link{createDT}()},
\code{\link{fillNA_CS_PP}()},
\code{\link{get_os}()},
\code{\link{is_empty}()}
}
\concept{general}
