% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_os.R
\name{get_os}
\alias{get_os}
\title{Get OS}
\usage{
get_os()
}
\value{
OS name.
}
\description{
Identify your current operating system (OS).
}
\examples{
os <- get_os()
}
\seealso{
Other general: 
\code{\link{check_if_empty}()},
\code{\link{createDT}()},
\code{\link{fillNA_CS_PP}()},
\code{\link{get_nrows}()}
}
\concept{general}
