% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_size.R
\name{get_sample_size}
\alias{get_sample_size}
\title{Get sample size}
\usage{
get_sample_size(
  dat,
  compute_n = c("ldsc", "giant", "metal", "sum"),
  return_only = NULL,
  force_new = FALSE,
  standardise_headers = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{dat}{Fine-mapping results data.}

\item{compute_n}{How to compute per-SNP sample size (new column "N").
\itemize{
\item{\code{0}: }{N will not be computed.}
\item{\code{>0}: }{If any number >0 is provided,
that value will be set as N for every row.
\strong{Note}: Computing N this way is incorrect and should be avoided
if at all possible.}
\item{\code{"sum"}: }{N will be computed as:
cases (N_CAS) + controls (N_CON), so long as both columns are present}.
\item{\code{"effective"}: }{N will be computed as effective sample size:
cases (N_CAS) + controls (N_CON), so long as both columns are present}.
}}

\item{return_only}{A function to return only a single value from the 
inferred/imputed sample size column (e.g. \code{max}, \code{min}).}

\item{force_new}{If "Neff" (or "N") already exists in \code{sumstats_dt},
replace it with the recomputed version.}

\item{standardise_headers}{Standardise headers first.}

\item{verbose}{Print messages.}

\item{...}{Additional argument passed to \code{return_only} function, 
if \code{return_only}  is not \code{NULL}.}
}
\description{
Infer sample size from summary stats using \pkg{MungeSumstats}.
}
\examples{
dat <- echodata::BST1
dat2 <- echodata::get_sample_size(dat = dat)
}
