% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_find_pages.R
\name{github_find_pages}
\alias{github_find_pages}
\title{Find pages}
\usage{
github_find_pages(
  creator = "RajLabMSSM",
  repo = "Fine_Mapping_Shiny",
  branch = "master",
  local_repo = NULL,
  return_table = TRUE,
  save_path = NULL
)
}
\arguments{
\item{creator}{Repo creator name
(can be organization or individual GitHub account name).}

\item{repo}{GitHub repository name.}

\item{branch}{Which branch to search.}
}
\description{
Find html files on GitHub Pages.
}
\keyword{internal}
