% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_list_files.R
\name{github_list_files}
\alias{github_list_files}
\title{List files in GitHub repo}
\usage{
github_list_files(
  creator = "RajLabMSSM",
  repo = "Fine_Mapping_Shiny",
  branch = c("main", "master"),
  query = NULL,
  return_download_api = TRUE,
  seconds = 5 * 60,
  verbose = TRUE
)
}
\arguments{
\item{creator}{Repo creator name
(can be organization or individual GitHub account name).}

\item{repo}{GitHub repository name.}

\item{branch}{Which branch to search.}

\item{query}{Query substring.}

\item{return_download_api}{Return the link to download each file
(instead of its path).}

\item{seconds}{number of seconds to wait for a response until giving up.
Can not be less than 1 ms.}

\item{verbose}{Print messages.}
}
\value{
A list of paths.
}
\description{
Search for files within a public GitHub repository and return their paths.
}
