% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_topSNPs.R
\name{import_topSNPs}
\alias{import_topSNPs}
\title{Import top GWAS/QTL summary statistics}
\usage{
import_topSNPs(
  topSS,
  show_table = TRUE,
  sheet = 1,
  munge = TRUE,
  chrom_col = NULL,
  position_col = NULL,
  min_POS_col = NULL,
  max_POS_col = NULL,
  snp_col = NULL,
  pval_col = NULL,
  effect_col = NULL,
  locus_col = "Locus",
  gene_col = "Gene",
  grouping_vars = c("Locus"),
  remove_variants = NULL,
  verbose = TRUE
)
}
\arguments{
\item{topSS}{Can be a data.frame with the top summary stats per locus.
Alternatively, you can provide a path to the stored top summary stats file.
Can be in any tabular format (e.g. excel, .tsv, .csv, etc.).
This file should have one lead GWAS/QTL hits per locus.
If there is more than one SNP per locus, the one with the
 smallest p-value (then the largest effect size) is selected as the lead SNP.
The lead SNP will be used as the center of the locus when
 constructing the locus subset files.}

\item{show_table}{Create an interative data table.}

\item{sheet}{If the \emph{topSS} file is an excel sheet, 
you can specify which tab to use.
You can provide either a number to identify the tab by order,
or a string to identify the tab by name.}

\item{munge}{Standardise column names.}

\item{chrom_col}{Column containing chromosome (can be "chr1" or "1" format).}

\item{position_col}{Column containing genomic position in units of basepairs.}

\item{min_POS_col}{Column containing minimum genomic position 
(used instead of an arbitrary window size).}

\item{max_POS_col}{Column containing maximum genomic position 
(used instead of an arbitrary window size).}

\item{snp_col}{Column containing SNP RSIDs.}

\item{pval_col}{Column containing uncorrected p-values.}

\item{effect_col}{Column containing effect size (e.g. Beta, OR).}

\item{locus_col}{Column containing unique locus name.}

\item{gene_col}{An optional column to keep track of the causal 
gene(s) in each locus (if known).}

\item{grouping_vars}{The variables that you want to group by
such that each grouping_var combination has its own index SNP.
For example, if you want one index SNP per QTL eGene - 
GWAS locus pair, you could supply:
\code{grouping_vars=c("Locus","Gene")}.}

\item{remove_variants}{SNPs to remove from \code{topSS},}

\item{verbose}{Print messages.}
}
\value{
Munged topSNPs table.
}
\description{
The resulting top_SNPs data.frame can be used to guide
the \link[echolocatoR]{finemap_loci} in querying and fine-mapping loci.
}
\examples{
topSNPs <- echodata::import_topSNPs(
    topSS = echodata::topSNPs_Nalls2019_raw, 
    pval_col = "P, all studies",
    effect_col = "Beta, all studies",
    locus_col = "Nearest Gene", 
    gene_col = "QTL Nominated Gene (nearest QTL)")
}
