% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_empty.R
\name{is_empty}
\alias{is_empty}
\title{Check if a file is empty}
\usage{
is_empty(path, error = TRUE, verbose = TRUE)
}
\arguments{
\item{path}{File path.}

\item{error}{Throw an error if the file is empty.}

\item{verbose}{Print messages.}
}
\value{
Boolean.
}
\description{
Check if a summary statistics file subset is empty.
}
\examples{
tmp <- tempfile()
data.table::fwrite(echodata::BST1, tmp)
empty <- echodata::is_empty(path = tmp)
}
\seealso{
Other general: 
\code{\link{createDT}()},
\code{\link{fillNA_CS_PP}()},
\code{\link{get_nrows}()},
\code{\link{get_os}()}
}
\concept{general}
