% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_ggbio.R
\name{is_ggbio}
\alias{is_ggbio}
\title{Is an object of class ggbio}
\usage{
is_ggbio(obj)
}
\arguments{
\item{obj}{Any R object.}
}
\value{
Boolean.
}
\description{
Check whether an object is of the class \link[ggbio]{ggbio}.
}
\examples{
bool <- echodata::is_ggbio(mtcars)
}
