% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_local.R
\name{is_local}
\alias{is_local}
\title{Is local}
\usage{
is_local(path, protocols = c("http", "https", "ftp", "ftps", "fttp", "fttps"))
}
\arguments{
\item{path}{Path to local file or remote URL.}

\item{protocols}{URL protocols to search for.}
}
\description{
Determine whether a file exist locally.
}
\examples{
echodata::is_local(
    path="https://github.com/RajLabMSSM/echotabix/raw/main/README.md"
)
}
