% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_url.R
\name{is_url}
\alias{is_url}
\title{Check if the input is url e.g. http:// or ftp://}
\source{
\href{https://rdrr.io/cran/seqminer/src/R/seqminer.R}{
Borrowed from \code{seqminer} internal function}
}
\usage{
is_url(path, protocols = c("http", "https", "ftp", "ftps", "fttp", "fttps"))
}
\arguments{
\item{path}{Path to local file or remote URL.}

\item{protocols}{URL protocols to search for.}
}
\description{
Check if the input is url e.g. http:// or ftp://
}
\examples{
path <- "https://github.com/RajLabMSSM/echolocatoR/raw/master/NEWS.md"
out <- is_url(path=path)
}
