% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{locus_dir}
\alias{locus_dir}
\title{Example results path for \code{Nall2019} BST1 locus}
\format{
path string
}
\source{
\code{
locus_dir <- "results/GWAS/Nalls23andMe_2019/BST1"
usethis::use_data(locus_dir, overwrite = TRUE)
}
}
\usage{
data("locus_dir")
}
\description{
Example results path for \code{Nall2019} BST1 locus
}
\keyword{datasets}
