% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LRRK2}
\alias{LRRK2}
\title{\pkg{echolocatoR} output example: LRRK2 locus}
\format{
data.table
\describe{
  \item{SNP}{SNP RSID}
  \item{CHR}{Chromosome}
  \item{POS}{Genomic position (in basepairs)}
  \item{...}{Optional: extra columns}
}
}
\source{
\url{https://doi.org/10.1016/S1474-4422(19)30320-5}
}
\usage{
data("LRRK2")
}
\description{
An example results file after running
\link[echolocatoR]{finemap_loci}.
}
\details{
Data originally comes from the Parkinson's disease GWAS
by \href{https://doi.org/10.1016/S1474-4422(19)30320-5}{
Nalls et al. (The Lancet Neurology)}.
}
\examples{
\dontrun{
library(echodata)
data("Nalls2019_merged")
LRRK2 <- subset(Nalls2019_merged, Locus=="LRRK2")
LRRK2 <- echodata::assign_lead_snp(LRRK2)
usethis::use_data(LRRK2, overwrite = TRUE)
}
}
\keyword{datasets}
