% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melt_finemapping_results.R
\name{melt_finemapping_results}
\alias{melt_finemapping_results}
\title{Melt fine-mapping results}
\usage{
melt_finemapping_results(dat, verbose = TRUE)
}
\arguments{
\item{dat}{Fine-mapping results data.}

\item{verbose}{Print messages.}
}
\value{
Melted \link[data.table]{data.table}
}
\description{
Melt fine-mapping results into long format such that 
"\*.CS" and "\*.PP" columns from multiple methods are now in a single column.
}
\examples{
finemap_melt <- echodata::melt_finemapping_results(dat = echodata::BST1)
}
