% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_robust.R
\name{merge_robust}
\alias{merge_robust}
\title{Merge robust}
\source{
\href{https://rdatatable.gitlab.io/data.table/articles/datatable-importing.html#optionally-import-data-table-suggests}{
data.table docs}
}
\usage{
merge_robust(x, y, by = NULL, dup_stop = TRUE, na_stop = TRUE, ...)
}
\arguments{
\item{x, y}{\code{data table}s. \code{y} is coerced to a \code{data.table} if
it isn't one already.}

\item{by}{A vector of shared column names in \code{x} and \code{y} to merge on.
This defaults to the shared key columns between the two tables.
If \code{y} has no key columns, this defaults to the key of \code{x}.}

\item{dup_stop}{Throw an error if duplicates of the \code{by} 
column are detected after merging.}

\item{na_stop}{Throw an error if all of the data from \code{y} 
has been converted to NAs (unless they were all NAs before merging).}

\item{...}{Additional arguments passed to 
\link[data.table]{merge.data.table}.}
}
\description{
Ensure two \link[data.table]{data.table}  are merged properly.
Necessary due to inconsistencies with how \link[data.table]{data.table} 
are evaluated in different scopes.
}
\examples{
dat <- echodata::BST1
x <- dat[,c("SNP","CHR","POS")]
y <- dat[,c("SNP","Support","leadSNP")]
z <- merge_robust(x=x, y=y, by="SNP")
}
