% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MEX3C}
\alias{MEX3C}
\title{\pkg{echolocatoR} output example: MEX3C locus}
\format{
data.table
\describe{
  \item{SNP}{SNP RSID}
  \item{CHR}{Chromosome}
  \item{POS}{Genomic position (in basepairs)}
  \item{...}{Optional: extra columns}
}
}
\source{
\url{https://doi.org/10.1016/S1474-4422(19)30320-5}
}
\usage{
data("MEX3C")
}
\description{
An example results file after running
\link[echolocatoR]{finemap_loci}.
}
\details{
Data originally comes from the Parkinson's disease GWAS
by \href{https://doi.org/10.1016/S1474-4422(19)30320-5}{
Nalls et al. (The Lancet Neurology)}.
}
\examples{
\dontrun{
library(echodata)
data("Nalls2019_merged")
MEX3C <- subset(Nalls2019_merged, Locus=="MEX3C")
MEX3C <- echoannot::assign_lead_SNP(MEX3C)
usethis::use_data(MEX3C, overwrite = TRUE)
}
}
\keyword{datasets}
