% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MUNGESUMSTATS.to_echolocatoR.R
\name{MUNGESUMSTATS.to_echolocatoR}
\alias{MUNGESUMSTATS.to_echolocatoR}
\title{Convert from \pkg{MungeSumstats} to \pkg{echolocatoR} format}
\usage{
MUNGESUMSTATS.to_echolocatoR(dat)
}
\arguments{
\item{dat}{Data}
}
\description{
Convert column names between package-specific formats.
}
\examples{
 
dat <- dplyr::rename(echodata::BST1, BP=POS, FRQ=Freq, BETA=Effect)
dat_echoR <- MUNGESUMSTATS.to_echolocatoR(dat=dat) 
}
