% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungesumstats_check_syn.R
\name{mungesumstats_check_syn}
\alias{mungesumstats_check_syn}
\title{Check colname synonyms}
\usage{
mungesumstats_check_syn(dat, col_name = "P-value", col_type = "P")
}
\description{
Check whether a table has headers that can be mapped 
to an attribute in \pkg{MungeSumstats}
}
\keyword{internal}
