% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungesumstats_col_map.R
\name{mungesumstats_col_map}
\alias{mungesumstats_col_map}
\title{Column name mappings from \pkg{MungeSumstats} to \pkg{echolocatoR}}
\source{
\url{https://github.com/neurogenomics/MungeSumstats}
}
\usage{
mungesumstats_col_map()
}
\description{
Column name mappings from \pkg{MungeSumstats} to \pkg{echolocatoR}
}
