% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Nalls2019}
\alias{Nalls2019}
\title{Example GWAS summary statistics: Nalls2019}
\format{
path string
}
\source{
\href{https://doi.org/10.1016/S1474-4422(19)30320-5}{
Nalls et al. 2019, The Lancet Neurology}

\code{
path <- echolocatoR::example_fullSS()
Nalls2019 <- data.table::fread(path)

# meta <- MungeSumstats::find_sumstats(ids = "ieu_b_7")
# Nalls2019 <- MungeSumstats::import_sumstats(ids = meta$id[1])
path <- file.path("/Volumes/bms20/projects/neurogenomics-lab/live",
                  "GWAS_sumstats/OpenGWAS/ieu-b-7.tsv.gz")

usethis::use_data(Nalls2019, overwrite = TRUE)
}
}
\usage{
data("Nalls2019")
}
\description{
Downsampled GWAS summary statistics from
 Nalls et al. 2019 Parkinson's Disease GWAS (without 23andMe data).
}
\keyword{datasets}
