% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_loci.R
\name{order_loci}
\alias{order_loci}
\title{Order loci}
\usage{
order_loci(dat, by_ucs_size = FALSE, descending = TRUE, verbose = FALSE)
}
\arguments{
\item{dat}{Fine-mapped summary statistics data.}

\item{by_ucs_size}{Sort by Union Credible Set (UCS) size.}

\item{descending}{Sort \code{by_ucs_size} in descending order.}

\item{verbose}{Print messages.}
}
\description{
Order loci by UCS size, or alphabetically.
}
\examples{
merged_DT <- echodata::get_Nalls2019_merged()
merged_DT2 <- echodata::order_loci(dat=merged_DT)
}
