% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/portal_query.R
\name{portal_query}
\alias{portal_query}
\title{Search and download fine-mapping files}
\usage{
portal_query(
  dataset_types = NULL,
  datasets = NULL,
  phenotypes = NULL,
  loci = NULL,
  LD_panels = c("UKB", "1KGphase1", "1KGphase3"),
  file_types = c("multi_finemap", "LD", "plot"),
  results_dir = tempdir(),
  overwrite = FALSE,
  as_datatable = FALSE,
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{dataset_types}{Dataset types to search for
(e.g. "GWAS", "QTL").}

\item{datasets}{Dataset names to search for
(e.g. "Ripke_2014", "Wray_2018").}

\item{phenotypes}{Phenotypes to search for
(e.g. "Schizophrenia", "Major Depressive Disorder").}

\item{loci}{Locus names to search for
(e.g. "BST1", "CHRNB1", "LRRK2").}

\item{LD_panels}{LD panels to search for,
or results generated using different LD panels
(e.g. "UKB", "1KGphase1", "1KGphase3").}

\item{file_types}{File types to search for:
\itemize{
\item{"multi_finemap":}{Fine-mapping results 
merged with GWAS/QTL summary statistics.}
\item{"LD":}{Linkage Disequilibrium with the lead SNP in each locus.}
\item{"plot" : }{Locus plots of fine-mapped results.}
}}

\item{results_dir}{Where to save the matching queries.}

\item{overwrite}{Whether to overwrite previously saved queries
with the same names.}

\item{as_datatable}{Return results organized in a 
\link[data.table]{data.table}. 
Otherwise, will return a list of local file paths.}

\item{nThread}{Number of threads to parallelise downloads across.}

\item{verbose}{Print messages.}
}
\value{
List of local paths where the requested files were downloaded to.
}
\description{
Search the \href{https://github.com/RajLabMSSM/Fine_Mapping_Shiny}{
echolocatoR Fine-mapping Portal}
for fine-mapping results, LD, and locus plots.
}
\examples{
local_finemap <- portal_query(
    dataset_types = "GWAS",
    phenotypes = c("schizophrenia", "parkinson"),
    file_types = "multi_finemap",
    loci = c("BST1", "CHRNB1", "LRRK2"),
    LD_panels = "1KGphase3") 
}
