% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview.R
\name{preview}
\alias{preview}
\title{Preview formatted sum stats saved to disk}
\usage{
preview(path, nrows = 5L, nThread = 1L)
}
\arguments{
\item{path}{Path to local file or remote URL. 
Alternatively, can pass in a data.frame to be used directly.}

\item{nrows}{ The maximum number of rows to read. Unlike \code{read.table}, you do not need to set this to an estimate of the number of rows in the file for better speed because that is already automatically determined by \code{fread} almost instantly using the large sample of lines. \code{nrows=0} returns the column names and typed empty columns determined by the large sample; useful for a dry run of a large file or to quickly check format consistency of a set of files before starting to read any of them. }

\item{nThread}{The number of threads to use. Experiment to see what works best for your data on your hardware.}
}
\value{
No return
}
\description{
Prints the first \code{nrows} lines of the sum stats.
}
