% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bgz.R
\name{read_bgz}
\alias{read_bgz}
\title{Read bgz}
\usage{
read_bgz(file, nrows = -1L, nThread = 1, verbose = TRUE, ...)
}
\arguments{
\item{file}{ File name in working directory, path to file (passed through \code{\link[base]{path.expand}} for convenience), or a URL starting http://, file://, etc. Compressed files with extension \file{.gz} and \file{.bz2} are supported if the \code{R.utils} package is installed. }

\item{nrows}{ The maximum number of rows to read. Unlike \code{read.table}, you do not need to set this to an estimate of the number of rows in the file for better speed because that is already automatically determined by \code{fread} almost instantly using the large sample of lines. \code{nrows=0} returns the column names and typed empty columns determined by the large sample; useful for a dry run of a large file or to quickly check format consistency of a set of files before starting to read any of them. }

\item{nThread}{The number of threads to use. Experiment to see what works best for your data on your hardware.}

\item{verbose}{ Be chatty and report timings? }

\item{...}{Additional arguments passed to \link[data.table]{fread}.}
}
\description{
Read a compressed bgz file into R.
}
\examples{
tmp <- tempfile(fileext = ".tsv.gz")
data.table::fwrite(echodata::BST1, file = tmp, sep = "\t")
file <- Rsamtools::bgzip(file = tmp, overwrite=TRUE)
dat <- echodata::read_bgz(file=file)
}
