% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_parquet.R
\name{read_parquet}
\alias{read_parquet}
\title{Read parquet file}
\usage{
read_parquet(path, verbose = TRUE)
}
\arguments{
\item{path}{Path to parquet file.}

\item{verbose}{Print messages.}
}
\description{
Import a file in parquet format as a
\link[data.table]{data.table}.
}
\examples{
 
#### Write a parquet file ####
dat <- echodata::BST1
path <- echodata::write_parquet(dat)
#### Read the parquet file back into R ####
dat2 <- echodata::read_parquet(path=path)
}
