% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_report.R
\name{results_report}
\alias{results_report}
\title{Results report}
\usage{
results_report(merged_dat)
}
\arguments{
\item{merged_dat}{Merged fine-mapping results generated by 
\link[echodata]{merge_finemapping_results}.}
}
\value{
Null
}
\description{
Generate a quick summary report of the fine-mapping results in aggregate.
}
\examples{
results_report(merged_dat = echodata::get_Nalls2019_merged()) 
}
\seealso{
Other summarise: 
\code{\link{get_CS_bins}()},
\code{\link{get_CS_counts}()},
\code{\link{get_SNPgroup_counts}()}
}
\concept{summarise}
