% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_group_filters.R
\name{snp_group_filters}
\alias{snp_group_filters}
\title{Get SNP group filters}
\usage{
snp_group_filters(invert = FALSE, random_sample_size = 20)
}
\arguments{
\item{invert}{Reverse the list.}

\item{random_sample_size}{Sample size for random SNP subset ("Random").}
}
\description{
Returns a character vector of filtering arguments for different SNP groups.
}
\examples{
filts <- echodata::snp_group_filters()
}
