% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize}
\alias{standardize}
\title{Standardize the locus subset}
\usage{
standardize(
  query,
  subset_path = NULL,
  locus = NULL,
  colmap = construct_colmap(),
  return_dt = TRUE,
  nThread = 1,
  verbose = TRUE
)
}
\description{
After querying a subset of the full summary statistics,
this function converts it into a standardized format
that the rest of \emph{echolocatoR} can work with.
}
\examples{
BST1 <- echodata::BST1
#### Screw up Freq to see if function can fix it and infer MAF ####
BST1$rsid <- BST1$SNP
BST1 <- data.frame(BST1)[,!colnames(BST1) \%in\% c("MAF","SNP")]
BST1[c(10,30,55),"Freq"] <- 0
BST1[c(12,22),"Freq"] <- NA

subset_path <- file.path(tempdir(),"BST1.tsv")
data.table::fwrite(BST1, subset_path)
query_mod <- echodata::standardize(subset_path = subset_path,
                                   locus = "BST1")
}
