% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_gene.R
\name{standardize_gene}
\alias{standardize_gene}
\title{Standardize genes}
\usage{
standardize_gene(dat, Gene = "gene", verbose = TRUE)
}
\arguments{
\item{dat}{GWAS/QTL summary statistics data.}

\item{Gene}{Name of the gene column in \code{dat}.}

\item{verbose}{Print messages.}
}
\description{
Convert ensembl IDs to gene symbols.
}
