% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_gene_locus_cols.R
\name{standardize_gene_locus_cols}
\alias{standardize_gene_locus_cols}
\title{Standardizse gene/locus columns}
\usage{
standardize_gene_locus_cols(
  topSNPs,
  Locus = "Locus",
  Gene = "Gene",
  grouping_vars = Locus,
  verbose = TRUE
)
}
\arguments{
\item{Locus}{Column containing unique locus name.}

\item{grouping_vars}{The variables that you want to group by
such that each grouping_var combination has its own index SNP.
For example, if you want one index SNP per QTL eGene - 
GWAS locus pair, you could supply:
\code{grouping_vars=c("Locus","Gene")}.}

\item{verbose}{Print messages.}
}
\value{
topSNPs
}
\description{
Standardizse Gene and Locus columns in \code{topSNPs}.
}
\keyword{internal}
