% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_maf.R
\name{standardize_maf}
\alias{standardize_maf}
\title{Standardize MAF}
\usage{
standardize_maf(query, colmap, verbose = TRUE)
}
\description{
Standardize Minor Allele Frequency column.
Will infer MAF from frequency column, assuming that 
MAF is always less than 0.5 and that all SNPs are biallelic
}
\keyword{internal}
