% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{standardize_subset}
\alias{standardize_subset}
\title{Standardize the locus subset}
\usage{
standardize_subset(
  subset_path,
  locus,
  top_SNPs = NULL,
  fullSS_genome_build = "hg19",
  chrom_col = "CHR",
  position_col = "POS",
  snp_col = "SNP",
  pval_col = "P",
  effect_col = "Effect",
  stderr_col = "StdErr",
  tstat_col = "t_stat",
  MAF_col = "MAF",
  freq_col = "Freq",
  N_cases_col = "N_cases",
  N_controls_col = "N_controls",
  N_cases = NULL,
  N_controls = NULL,
  proportion_cases = "calculate",
  sample_size = NULL,
  A1_col = "A1",
  A2_col = "A2",
  gene_col = "Gene",
  qtl_prefixes = NULL,
  return_dt = TRUE,
  nThread = 1,
  download_method = "axel",
  verbose = TRUE
)
}
\description{
After querying a subset of the full summary statistics,
this function converts it into a standardized format
that the rest of \emph{echolocatoR} can work with.
}
\examples{
BST1 <- echodata::BST1
#### Screw up Freq to see if function can fix it and infer MAF ####
BST1$rsid <- BST1$SNP
BST1 <- data.frame(BST1)[,!colnames(BST1) \%in\% c("MAF","SNP")]
BST1[c(10,30,55),"Freq"] <- 0
BST1[c(12,22),"Freq"] <- NA

subset_path <- file.path(tempdir(),"BST1.tsv")
data.table::fwrite(BST1, subset_path)
query_mod <- echolocatoR:::standardize_subset(subset_path=subset_path,
                                              locus="BST1",
                                              snp_col = "rsid",
                                              MAF_col="calculate")
}
\concept{standardizing functions}
