% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sumstatsColHeaders}
\alias{sumstatsColHeaders}
\title{\code{sumstatsColHeaders} from \pkg{MungeSumstats}}
\format{
An object of class \code{data.frame} with 205 rows and 2 columns.
}
\usage{
data("sumstatsColHeaders")
}
\description{
\code{sumstatsColHeaders} from \pkg{MungeSumstats}
}
\examples{
\dontrun{
sumstatsColHeaders <- MungeSumstats:::sumstatsColHeaders
usethis::use_data(sumstatsColHeaders, overwrite=TRUE)
}
}
\keyword{datasets}
