% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{top_SNPs}
\alias{top_SNPs}
\title{TopSS example file (processed)}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 97 rows and 7 columns.
}
\source{
\code{
data("Nalls_top_SNPs")
top_SNPs <- echolocatoR::import_topSNPs(topSS=Nalls_top_SNPs,
                                        chrom_col="CHR",
                                        position_col="BP", 
                                        snp_col="SNP", 
                                        pval_col="P, all studies", 
                                        effect_col="Beta, all studies", 
                                        gene_col="Nearest Gene",  
                                        locus_col="Nearest Gene", 
                                        remove_variants="rs34637584")
usethis::use_data(top_SNPs, overwrite=TRUE)
}
}
\usage{
top_SNPs
}
\description{
Summary stats of the top SNP(s) per locus.
Used to query locus subsets.for fine-mapping.
}
\keyword{datasets}
