% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{topSNPs_Kunkle2019}
\alias{topSNPs_Kunkle2019}
\title{TopSS example file (processed): Kunkle2019}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 8 rows and 7 columns.
}
\source{
\code{
library(dplyr)
path <- file.path("/Volumes/bms20/projects/neurogenomics-lab/live",
                  "GWAS_sumstats/OpenGWAS/ieu-b-2.tsv.gz")
fullSS <- data.table::fread(path)

supp_remote <- file.path(
    "https://static-content.springer.com/esm",
    "art%3A10.1038%2Fs41588-019-0358-2",
    "MediaObjects/41588_2019_358_MOESM3_ESM.xlsx")
supp_local <- file.path(tempdir(),basename(supp_remote))
utils::download.file(supp_remote, supp_local)
topSS <- readxl::read_excel(supp_local,
                            sheet = "Supplementary Table 8",
                            skip = 2) %>%
    dplyr::rename(SNP="Top Associated SNV")
topSS <- merge(topSS,
      subset(fullSS, SNP %in% topSS$SNP),
      by="SNP")

paths <- googledrive::drive_download(
    file.path("https://docs.google.com/spreadsheets/d",
              "1BgLQaRZd9L7JoO8IbpzhUCRFdTdLgJvD/edit#gid=236666677"),
    overwrite = TRUE)
meta <- readxl::read_excel(paths$local_path, sheet = "GWAS")
meta <- subset(meta, dataset=="Kunkle_2019")

top_SNPs <- echolocatoR::import_topSNPs(
    topSS = topSS,
    sheet = meta$top_sheet,
    chrom_col = meta$top_chrom,
    position_col = meta$top_pos,
    snp_col = meta$top_snp,
    pval_col = "P",
    effect_col = "BETA",
    locus_col = meta$top_locus)
topSNPs_Kunkle2019 <- top_SNPs
usethis::use_data(topSNPs_Kunkle2019, overwrite=TRUE)
}
}
\usage{
topSNPs_Kunkle2019
}
\description{
Summary stats of the top SNP(s) per locus.
Used to query locus subsets.for fine-mapping.
}
\keyword{datasets}
