% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{topSNPs_Nalls2019}
\alias{topSNPs_Nalls2019}
\title{TopSS example file (processed): Nalls2019}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 97 rows and 7 columns.
}
\source{
\code{
topSS <- echodata::topSNPs_Nalls2019_raw
topSNPs_Nalls2019 <- echodata::import_topSNPs(topSS=topSS,
                                        chrom_col="CHR",
                                        position_col="BP",
                                        snp_col="SNP",
                                        pval_col="P, all studies",
                                        effect_col="Beta, all studies",
                                        gene_col="Nearest Gene",
                                        locus_col="Nearest Gene",
                                        remove_variants="rs34637584")
usethis::use_data(topSNPs_Nalls2019, overwrite=TRUE)
}
}
\usage{
data("topSNPs_Nalls2019")
}
\description{
Summary stats of the top SNP(s) per locus.
Used to query locus subsets.for fine-mapping.
}
\details{
Formerly \code{top_SNPs}.
}
\keyword{datasets}
