% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{topSNPs_Nalls2019_raw}
\alias{topSNPs_Nalls2019_raw}
\title{TopSS example file: Nalls2019}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 107 rows and 31 columns.
}
\source{
\url{https://github.com/RajLabMSSM/Fine_Mapping/raw/master/Data/GWAS/Nalls23andMe_2019/Nalls2019_TableS2.xlsx}
\code{
local <- file.path(tempdir(),"Nalls2019_TableS2.xlsx")
utils::download.file(
    file.path("https://github.com/RajLabMSSM/Fine_Mapping",
              "raw/master/Data/GWAS/Nalls23andMe_2019",
              "Nalls2019_TableS2.xlsx"),
    local)
topSNPs_Nalls2019_raw <- data.table::data.table(readxl::read_excel(local))
usethis::use_data(topSNPs_Nalls2019_raw, overwrite=TRUE)
}
}
\usage{
topSNPs_Nalls2019_raw
}
\description{
Summary stats of the top SNP(s) per locus.
Used to query locus subsets.for fine-mapping.
}
\details{
Formerly \code{topSNPs_Nalls2019}.

Data from \href{https://doi.org/10.1016/S1474-4422(19)30320-5}{
Nalls et al. (bioRxiv)}, Table S2.
}
\keyword{datasets}
