% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_cols.R
\name{update_cols}
\alias{update_cols}
\title{Update CS cols}
\usage{
update_cols(dat)
}
\arguments{
\item{dat}{Fine-mapping results data.}
}
\description{
Convert old column format: ".Credible_Set" => ".CS"
}
\examples{
dat <- echodata::update_cols(dat = echodata::BST1) 
}
