% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_parquet.R
\name{write_parquet}
\alias{write_parquet}
\title{Write parquet file}
\usage{
write_parquet(
  dat,
  path = tempfile(fileext = ".parquet"),
  conda_env = "echoR",
  conda = "auto",
  verbose = TRUE
)
}
\arguments{
\item{dat}{\link[base]{data.frame}.}

\item{path}{Path to write parquet file to.}

\item{conda_env}{Conda environment to use.}

\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary.
See \strong{Finding Conda} and \code{\link[reticulate:conda-tools]{conda_binary()}} for more details.}

\item{verbose}{Print messages.}
}
\description{
Export a \link[base]{data.frame} to parquet format (from python).
}
\examples{
dat <- echodata::BST1
path <- echodata::write_parquet(dat = dat)
}
