% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_lead_snp.R
\name{assign_lead_snp}
\alias{assign_lead_snp}
\title{Assign lead SNP}
\usage{
assign_lead_snp(dat, grouping_vars = "Locus", verbose = TRUE)
}
\arguments{
\item{dat}{Fine-mapping data.table.}

\item{verbose}{Print messages.}
}
\value{
Modified data.table.
}
\description{
Assign lead SNP by creating a new boolean column "leadSNP" 
indicating whether a SNP has the smallest p-value ("P).
If multiple rows have the same p-value, 
 the one with the largest effective size ("Effect") is used as the lead SNP.
}
\examples{
dat <- echodata::assign_lead_snp(dat = echodata::BST1)
}
