% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_if_empty.R
\name{check_if_empty}
\alias{check_if_empty}
\title{Check if a file is empty}
\usage{
check_if_empty(file_path)
}
\arguments{
\item{file_path}{File path.}
}
\value{
Boolean.
}
\description{
Check if a summary statistics file subset is empty.
}
\examples{
tmp <- tempfile()
data.table::fwrite(data.table::data.table(), tmp)

is_empty <- echodata::check_if_empty(file_path = tmp)
}
\seealso{
Other general: 
\code{\link{createDT}()},
\code{\link{fillNA_CS_PP}()},
\code{\link{get_header}()}
}
\concept{general}
