% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords_to_rsids.R
\name{coords_to_rsids}
\alias{coords_to_rsids}
\title{Get RSIDS from SNP coordinates}
\usage{
coords_to_rsids(
  dat,
  genome_build = "hg19",
  drop_unannotated = TRUE,
  drop_duplicates = TRUE,
  SNPname = "SNP",
  verbose = TRUE
)
}
\arguments{
\item{dat}{\link[data.table]{data.table} or 
\link[base]{data.frame} with SNP positions in the columns 
\code{CHR} and \code{POS}. Extra columns are allowed.}

\item{genome_build}{Which genome build \code{dat} is in
(e.g. "hg19" or "hg38").}

\item{drop_unannotated}{Drop SNPs that RSIDs couldn't be found for.}

\item{drop_duplicates}{Drop any duplicate SNPs rows.}

\item{SNPname}{Name of the new column with RSIDs.}

\item{verbose}{Print messages.}
}
\value{
\link[data.table]{data.table} with new columns for RSIDs.
}
\description{
Get RSIDS from SNP coordinates using Bioconductor genome references.
}
\examples{
dat <- echodata::BST1[seq(5),]
data.table::setnames(dat,"SNP","SNP_old")
dat_annot <- coords_to_rsids(dat = dat)
}
