% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createDT.R
\name{createDT}
\alias{createDT}
\title{Interactive DT}
\usage{
createDT(dat, caption = "", scrollY = 400)
}
\arguments{
\item{dat}{Data to show.}

\item{caption}{Table caption.}

\item{scrollY}{Maximum height (in pixels) before defaulting to scrolling.}
}
\description{
Generate an interactive data table with download buttons.
}
\examples{
 
echodata::createDT(dat = echodata::BST1)
}
\seealso{
Other general: 
\code{\link{check_if_empty}()},
\code{\link{fillNA_CS_PP}()},
\code{\link{get_nrows}()}
}
\concept{general}
