% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_loci.R
\name{extract_loci}
\alias{extract_loci}
\title{Extract loci from GWAS/QTL summary statistics}
\usage{
extract_loci(topSNPs, fullSS, bp_distance = 5e+05, munged = TRUE)
}
\arguments{
\item{topSNPs}{Positions of the top SNPs prepare using
\link[echolocatoR]{import_topSNPs}.}

\item{fullSS}{\link[data.table]{data.table} of the 
full GWAS/QTL summary statistics.}

\item{bp_distance}{The width of the window size you want each locus to be.
For example, if \code{bp_distance=500000} then the locus will span 500kb from the lead SNP in either direction,
resulting in a locus that is ~1Mb long (depending on the dataset).}

\item{munged}{Whether \code{fullSS} was previously munged with 
\link[MungeSumstats]{format_sumstats}.}
}
\description{
Extract loci from GWAS/QTL summary statistics
}
