% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_consensus_snps.R
\name{find_consensus_snps}
\alias{find_consensus_snps}
\title{Find Consensus SNPs}
\usage{
find_consensus_snps(
  dat,
  verbose = TRUE,
  credset_thresh = 0.95,
  consensus_thresh = 2,
  sort_by_support = TRUE,
  exclude_methods = NULL
)
}
\description{
Find Consensus SNPs in \link[echolocatoR]{finemap_loci} output.
}
\examples{
BST1 <- echodata::find_consensus_snps(dat = echodata::BST1)
}
\seealso{
Other echolocatoR: 
\code{\link{find_consensus_snps_no_polyfun}()}
}
\concept{echolocatoR}
