% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_header.R
\name{get_header}
\alias{get_header}
\title{Get file header}
\usage{
get_header(path, colnames_only = TRUE, nrows = 2L, nThread = 1L)
}
\arguments{
\item{path}{Path to local file or remote URL.}

\item{colnames_only}{Whether to return only column names 
(default: \code{TRUE}), or the first \code{n} rows as well.}

\item{nrows}{ The maximum number of rows to read. Unlike \code{read.table}, you do not need to set this to an estimate of the number of rows in the file for better speed because that is already automatically determined by \code{fread} almost instantly using the large sample of lines. \code{nrows=0} returns the column names and typed empty columns determined by the large sample; useful for a dry run of a large file or to quickly check format consistency of a set of files before starting to read any of them. }

\item{nThread}{The number of threads to use. Experiment to see what works best for your data on your hardware.}
}
\description{
Reading in this way is more robust and able to handle bgz format.
}
\examples{
tmp <- tempfile(fileext = ".tsv.gz")
data.table::fwrite(echodata::BST1, file = tmp, sep = "\t") 
header <- echodata::get_header(path=tmp)
}
