% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_header.R
\name{get_header}
\alias{get_header}
\title{Get file header}
\usage{
get_header(large_file, colnames_only = TRUE, n = 2, nThread = 1)
}
\description{
Reading in this way is more robust and able to handle bgz format.
}
\seealso{
Other general: 
\code{\link{check_if_empty}()},
\code{\link{createDT}()},
\code{\link{fillNA_CS_PP}()},
\code{\link{get_sample_size}()}
}
\concept{general}
\keyword{internal}
