% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_download_files.R
\name{github_download_files}
\alias{github_download_files}
\title{Download files from GitHub}
\usage{
github_download_files(
  filelist,
  download_dir = tempdir(),
  overwrite = FALSE,
  timeout = 5 * 60,
  nThread = 1,
  verbose = TRUE
)
}
\description{
Download files from GitHub
}
\keyword{internal}
