% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/granges_to_bed.R
\name{granges_to_bed}
\alias{granges_to_bed}
\title{GRanges to BED format}
\usage{
granges_to_bed(
  GR.annotations,
  output_path,
  sep = "\\t",
  nThread = 1,
  gzip = FALSE
)
}
\description{
Convert GRanges object to BED format and save
}
\keyword{internal}
