% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Kunkle2019}
\alias{Kunkle2019}
\title{Example GWAS summary statistics}
\format{
path string
}
\source{
\code{
# meta <- MungeSumstats::find_sumstats(ids = "ieu_b_2")
# Kunkle2019 <- MungeSumstats::import_sumstats(ids = meta$id[1])
#path <- file.path("/Volumes/bms20/projects/neurogenomics-lab/live",
#                  "GWAS_sumstats/OpenGWAS/ieu-b-2.tsv.gz")
path <- "~/Desktop/ieu-b-2.tsv.gz"
Kunkle2019 <- data.table::fread(path)
Kunkle2019 <- subset(Kunkle2019, P<5e-8)
usethis::use_data(Kunkle2019, overwrite = TRUE)
}
}
\usage{
data("Kunkle2019")
}
\description{
Downsampled GWAS summary statistics from
 Kunkle et al. 2019 Alzheimer's Disease GWAS.
}
\keyword{datasets}
