% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limit_snps.R
\name{limit_snps}
\alias{limit_snps}
\title{Limit the number of SNPs per locus.}
\usage{
limit_snps(max_snps = 500, dat)
}
\arguments{
\item{max_snps}{The maximum number of SNPs to keep in the resulting data.frame.}

\item{dat}{A data.frame that contains at least the following columns:
\describe{
  \item{SNP}{RSID for each SNP.}
  \item{POS}{Each SNP's genomic position (in basepairs).}
}}
}
\description{
Start with the lead SNP and keep expanding the window until you reach the desired number of snps.
\code{dat} should only contain one locus from one chromosome.
}
\seealso{
Other SNP filters: 
\code{\link{filter_snps}()},
\code{\link{gene_trimmer}()}
}
\concept{SNP filters}
\keyword{internal}
