% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Nalls2019}
\alias{Nalls2019}
\title{Example GWAS summary statistics}
\format{
path string
}
\source{
\code{
path <- echolocatoR::example_fullSS()
Nalls2019 <- data.table::fread(path)
usethis::use_data(Nalls2019, overwrite = TRUE)
}
}
\usage{
data("Nalls2019")
}
\description{
Downsampled GWAS summary statistics from
 Nalls et al. 2019 Parkinson's Disease GWAS (without 23andMe data).
}
\keyword{datasets}
