% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_parquet.R
\name{read_parquet}
\alias{read_parquet}
\title{Read parquet file}
\usage{
read_parquet(path, conda_env = "echoR", conda = "auto", verbose = TRUE)
}
\arguments{
\item{path}{Path to parquet file.}

\item{conda_env}{Conda environment to use.}

\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary.
See \strong{Finding Conda} and \code{\link[reticulate:conda-tools]{conda_binary()}} for more details.}

\item{verbose}{Print messages.}
}
\description{
Import a file in parquet format (from python) as a
\link[data.table]{data.table} in R.
}
\examples{
 
#### Write a parquet file ####
dat <- echodata::BST1
path <- echodata::write_parquet(dat)
#### Read the parquet file back into R ####
dat2 <- echodata::read_parquet(path=path)
}
