% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_parquet.R
\name{read_parquet}
\alias{read_parquet}
\title{Read parquet file}
\usage{
read_parquet(
  path,
  conda_env = "echoR",
  conda = "auto",
  method = "pandas",
  verbose = TRUE
)
}
\arguments{
\item{path}{Path to parquet file.}

\item{conda}{The path to a \code{conda} executable. Use \code{"auto"} to allow
\code{reticulate} to automatically find an appropriate \code{conda} binary.
See \strong{Finding Conda} and \code{\link[reticulate:conda-tools]{conda_binary()}} for more details.}

\item{method}{Method to read parquet with:
\itemize{
\item{"pandas"}{Python package,
 but can be imported via \link[reticulate]{import.}}
\item{"SparkR"}{R package,
 but requires installation of SQL first.}
}}

\item{verbose}{Print messages.}
}
\description{
Import a file in parquet format (from python) as a
\link[data.table]{data.table} in R.
}
\examples{
path <- system.file(
    "tools/polyfun/example_data/weights.10.l2.ldscore.parquet",
    package = "echofinemap")
parq <- echofinemap::read_parquet(path=path)
}
