% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_parquet.R
\name{read_parquet}
\alias{read_parquet}
\title{Read parquet file}
\usage{
read_parquet(path, verbose = TRUE, ...)
}
\arguments{
\item{path}{Path to parquet file.}

\item{verbose}{Print messages.}

\item{...}{
  Arguments passed on to \code{\link[arrow:read_parquet]{arrow::read_parquet}}
  \describe{
    \item{\code{file}}{A character file name or URI, \code{raw} vector, an Arrow input stream,
or a \code{FileSystem} with path (\code{SubTreeFileSystem}).
If a file name or URI, an Arrow \link[arrow]{InputStream} will be opened and
closed when finished. If an input stream is provided, it will be left
open.}
    \item{\code{col_select}}{A character vector of column names to keep, as in the
"select" argument to \code{data.table::fread()}, or a
\link[tidyselect:eval_select]{tidy selection specification}
of columns, as used in \code{dplyr::select()}.}
    \item{\code{as_data_frame}}{Should the function return a \code{data.frame} (default) or
an Arrow \link[arrow]{Table}?}
    \item{\code{props}}{\link[arrow]{ParquetArrowReaderProperties}}
  }}
}
\description{
Import a file in parquet format as a
\link[data.table]{data.table}.
}
\examples{
 
#### Write a parquet file ####
dat <- echodata::BST1
path <- echodata::write_parquet(dat)
#### Read the parquet file back into R ####
dat2 <- echodata::read_parquet(path=path)
}
