% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_report.R
\name{results_report}
\alias{results_report}
\title{Give a quick summary report of the fine-mapping results}
\usage{
results_report(merged_dat)
}
\value{
Null
}
\description{
Give a quick summary report of the fine-mapping results
}
\examples{
results_report(echodata::get_Nalls2019_merged()) 
}
\seealso{
Other summarise: 
\code{\link{get_CS_bins}()},
\code{\link{get_CS_counts}()},
\code{\link{get_SNPgroup_counts}()}
}
\concept{summarise}
