% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_parquet.R
\name{write_parquet}
\alias{write_parquet}
\title{Write parquet file}
\usage{
write_parquet(dat, path = tempfile(fileext = ".parquet"), verbose = TRUE)
}
\arguments{
\item{dat}{\link[base]{data.frame}.}

\item{path}{Path to write parquet file to.}

\item{verbose}{Print messages.}
}
\description{
Export a \link[base]{data.frame} to parquet format.
}
\examples{
dat <- echodata::BST1
path <- echodata::write_parquet(dat = dat) 
}
